+++ Sources/Surge/Auxiliary Functions/Auxiliary.swift
