+++ Sources/Surge/Linear Algebra/Scalar.swift
